test(format): add and split additional tests
